removes abortNotFound condition from patching
this is not a reliable indicator of whether or not the object still
exists. during a "finalization", the object will be deleted after the
object is patched to remove the annotation resulting in an error when
attempting to subsequently patch that same object's status.

furthermore, when reconciling delete events directly, we need to ignore
not found in both cases since the object will not exist in that loop